Moving calls into routes from Bridge view to separate hooks
---------
