+++ wormhole-connect/src/views/Bridge/RouteOptions.tsx
@@ -13,7 +13,7 @@ import { calculateUSDPrice, getDisplayName } from 'utils';
 import config from 'config';
 import { Route } from 'config/types';
 import { RoutesConfig, RouteData } from 'config/routes';
-import { ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
+import type { ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
 
 import BridgeCollapse, { CollapseControlStyle } from './Collapse';
 import TokenIcon from 'icons/TokenIcons';
