+++ wormhole-connect/src/views/Redeem/GovernorEnqueuedWarning.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { makeStyles } from 'tss-react/mui';
-import { ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
+import type { ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
 import config from 'config';
 import { GOVERNOR_WHITEPAPER_URL } from 'consts';
 import AlertBanner from 'components/AlertBanner';
